requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.15.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.3
certifi==2023.5.7
charset-normalizer==3.1.0
discord.py==2.3.1
dnspython==1.16.0
frozenlist==1.3.3
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.1
pluggy==1.2.0
proto-plus==1.23.0
protobuf==4.25.3
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==3.12.0
pyparsing==3.1.2
pytest==7.1.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
soupsieve==2.4.1
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.0.3
yarl==1.9.2
